test(server): add helper for creating api functions
